+++ tailwindcss-elevation/.eslintrc.js
@@ -3,15 +3,15 @@ module.exports = {
     commonjs: true,
     es6: true,
     mocha: true,
-    node: true
+    node: true,
   },
   extends: ["standard", "plugin:prettier/recommended"],
   globals: {
     Atomics: "readonly",
-    SharedArrayBuffer: "readonly"
+    SharedArrayBuffer: "readonly",
   },
   parserOptions: {
-    ecmaVersion: 2018
+    ecmaVersion: 2018,
   },
-  rules: {}
+  rules: {},
 };