+++ tailwindcss-elevation/test/fixtures/hex-color-config.js
@@ -1,3 +1,4 @@
 module.exports = {
-  plugins: [require("../../index")([], { color: "#4FD1C5" })]
+  content: ["./test/fixtures/example.html"],
+  plugins: [require("../../index")({ color: "#4FD1C5" })],
 };