+++ tailwindcss-elevation/test/fixtures/invalid-config.js
@@ -1,5 +1,6 @@
 module.exports = {
+  content: ["./test/fixtures/example.html"],
   plugins: [
-    require("../../index")([], { color: "invalid", opacityBoost: "invalid" })
-  ]
+    require("../../index")({ color: "invalid", opacityBoost: "invalid" }),
+  ],
 };