+++ tailwindcss-elevation/test/fixtures/var-color-config.js
@@ -1,3 +1,4 @@
 module.exports = {
-  plugins: [require("../../index")([], { color: "var(--color)" })]
+  content: ["./test/fixtures/example.html"],
+  plugins: [require("../../index")({ color: "var(--color)" })],
 };