+++ tailwindcss-elevation/test/fixtures/variants-config.js
@@ -1,3 +1,0 @@
-module.exports = {
-  plugins: [require("../../index")(["responsive"])]
-};