+++ tailwindcss-elevation/test/utilities.test.js
@@ -2,22 +2,22 @@ const assert = require("chai").assert;
 
 const utilities = require("../src/utilities");
 
-describe("#utilities()", function() {
-  it("should return a utility for .elevation-0", function() {
+describe("#utilities()", function () {
+  it("should return a utility for .elevation-0", function () {
     assert.equal(
       utilities()[".elevation-0"].boxShadow,
       "0px 0px 0px 0px rgba(0,0,0,0.20), 0px 0px 0px 0px rgba(0,0,0,0.14), 0px 0px 0px 0px rgba(0,0,0,0.12)"
     );
   });
 
-  it("should return a utility for .elevation-24", function() {
+  it("should return a utility for .elevation-24", function () {
     assert.equal(
       utilities()[".elevation-24"].boxShadow,
       "0px 11px 15px -7px rgba(0,0,0,0.20), 0px 24px 38px 3px rgba(0,0,0,0.14), 0px 9px 46px 8px rgba(0,0,0,0.12)"
     );
   });
 
-  it("should return a utility for .elevation-1 when config includes color with an RGB triplet", function() {
+  it("should return a utility for .elevation-1 when config includes color with an RGB triplet", function () {
     const config = { color: "255,255,0" };
 
     assert.equal(
@@ -26,7 +26,7 @@ describe("#utilities()", function() {
     );
   });
 
-  it("should return a utility for .elevation-1 when config includes color with a HEX triplet", function() {
+  it("should return a utility for .elevation-1 when config includes color with a HEX triplet", function () {
     const config = { color: "#E53E3E" };
 
     assert.equal(
@@ -35,7 +35,7 @@ describe("#utilities()", function() {
     );
   });
 
-  it("should return a utility for .elevation-1 when config includes color with a custom property", function() {
+  it("should return a utility for .elevation-1 when config includes color with a custom property", function () {
     const config = { color: "var(--blue)" };
 
     assert.equal(
@@ -44,7 +44,7 @@ describe("#utilities()", function() {
     );
   });
 
-  it("should return a utility for .elevation-2 when config includes opacityBoost", function() {
+  it("should return a utility for .elevation-2 when config includes opacityBoost", function () {
     const config = { opacityBoost: ".23" };
     assert.equal(
       utilities(config)[".elevation-2"].boxShadow,
@@ -52,7 +52,7 @@ describe("#utilities()", function() {
     );
   });
 
-  it("should return a utility for .elevation-3 when config includes color and opacityBoost", function() {
+  it("should return a utility for .elevation-3 when config includes color and opacityBoost", function () {
     const config = { color: "77,192,181", opacityBoost: "0.42" };
 
     assert.equal(
@@ -61,7 +61,7 @@ describe("#utilities()", function() {
     );
   });
 
-  it("should return undefined for .elevation-25", function() {
+  it("should return undefined for .elevation-25", function () {
     assert.isUndefined(utilities()[".elevation-25"]);
   });
 });